feat(price-pusher): Further refactoring
