+++ stagecoach/price-pusher/price_pusher/main.py
@@ -1,5 +1,4 @@
 import click
-import os
 import logging
 
 from typing import Optional, Union
@@ -11,8 +10,8 @@ from pragma.publisher.client import (
     FetcherClient,
 )
 
+from price_pusher.utils.cli import setup_logging, load_private_key, create_client
 from price_pusher.configs import PriceConfig
-from price_pusher.utils.aws import fetch_aws_private_key
 
 
 logger = logging.getLogger(__name__)
@@ -93,8 +92,10 @@ def main(
             "API key and API URL are required when destination is 'offchain'."
         )
     # Build needed parameters
-    setup_logging(log_level)
+    setup_logging(logger, log_level)
     private_key = load_private_key(private_key)
+    _price_config = PriceConfig.from_yaml(config_file)
+    # Create & execute the client
     client: Union[PragmaOnChainClient, PragmaAPIClient] = create_client(
         target=target,
         network=network,
@@ -103,100 +104,10 @@ def main(
         api_key=api_key,
         api_url=api_url,
     )
-    price_config = PriceConfig.from_yaml(config_file)
-    print(price_config)
-    # Create & execute the client
     publisher_client = PragmaClient(client)
     fetcher_client = FetcherClient()
     # run_puller(target, net, keys, publisher)
     # run_pusher(target, net, keys, publisher)
-
-
-def create_client(
-    target: str,
-    network: str,
-    publisher_address: str,
-    private_key: str,
-    api_key: Optional[str] = None,
-    api_url: Optional[str] = None,
-) -> Union[PragmaOnChainClient, PragmaAPIClient]:
-    """
-    Create the appropriate client based on the target.
-
-    Args:
-        target: The target type ('onchain' or 'offchain').
-        network: The network type.
-        publisher_address: The publisher's address.
-        private_key: The private key for the account.
-        api_key: The API key for offchain publishing.
-        api_url: The API base URL for offchain publishing.
-
-    Returns:
-        Union[PragmaOnChainClient, PragmaAPIClient]
-    """
-    if target == "onchain":
-        return PragmaOnChainClient(
-            network=network,
-            account_contract_address=publisher_address,
-            account_private_key=private_key,
-        )
-    elif target == "offchain":
-        if not api_key:
-            raise click.BadParameter(
-                "Argument api-key can't be None if offchain is selected"
-            )
-        if not api_url:
-            raise click.BadParameter(
-                "Argument api-url can't be None if offchain is selected"
-            )
-        return PragmaAPIClient(
-            account_contract_address=publisher_address,
-            account_private_key=private_key,
-            api_key=api_key,
-            api_base_url=api_url,
-        )
-    else:
-        raise ValueError(f"Invalid target: {target}")
-
-
-def setup_logging(log_level: str) -> None:
-    """
-    Set up the logging configuration based on the provided log level.
-
-    Args:
-        log_level: The logging level to set (e.g., "DEBUG", "INFO").
-    """
-    numeric_log_level = getattr(logging, log_level.upper(), None)
-    if numeric_log_level is None:
-        raise ValueError(f"Invalid log level: {log_level}")
-
-    logging.basicConfig(level=numeric_log_level)
-    logger.setLevel(numeric_log_level)
-    logging.getLogger().setLevel(numeric_log_level)
-
-
-def load_private_key(private_key: str) -> str:
-    """
-    Load the private key either from AWS, environment variable, or from the provided plain value.
-
-    Args:
-        private_key: The private key string, either prefixed with 'aws:', 'plain:', or 'env:'.
-
-    Returns:
-        str
-    """
-    if private_key.startswith("aws:"):
-        secret_name = private_key.split("aws:", 1)[1]
-        return fetch_aws_private_key(secret_name)
-    elif private_key.startswith("plain:"):
-        return private_key.split("plain:", 1)[1]
-    elif private_key.startswith("env:"):
-        env_var_name = private_key.split("env:", 1)[1]
-        return os.getenv(env_var_name)
-    else:
-        raise click.UsageError(
-            "Private key must be prefixed with either 'aws:', 'plain:', or 'env:'"
-        )
 
 
 if __name__ == "__main__":
