+++ stagecoach/price-pusher/price_pusher/type.py
@@ -1,5 +1,8 @@
 from enum import Enum
 
+UnixTimestamp = int
+DurationInSeconds = int
+
 
 class Envirronment(Enum):
     DEV = 1
@@ -10,7 +13,3 @@ class DataSource(Enum):
     ONCHAIN = 1
     OFFCHAIN = 2
     DEFILLAMA = 3
-
-
-UnixTimestamp = int
-DurationInSeconds = int
